# apps/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hickory Docs API",
    "description": "",
    "license": {
      "name": "GPL-3.0-or-later",
      "identifier": "GPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/reset/confirm": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Complete a password reset.",
        "operationId": "confirm_reset",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "password reset",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/reset/request": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Begin a password reset.",
        "description": "Always answers 200, whether or not the address has an account. Anything\nelse turns this endpoint into an account-existence oracle, which is the\nclassic way a reset flow leaks the user list.",
        "operationId": "request_reset",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "reset email queued if the address exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/signup": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "signup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "account created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/verify/confirm": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Redeem a verification link.",
        "operationId": "confirm_verification",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "email verified",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusOut"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthOut": {
        "type": "object",
        "required": [
          "token",
          "user"
        ],
        "properties": {
          "token": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          }
        }
      },
...
      "Credentials": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
      "EmailBody": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string"
          }
        }
      },
...
      "ResetBody": {
        "type": "object",
        "required": [
          "token",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        }
      },
...
      "StatusOut": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
      "TokenBody": {
        "type": "object",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
...
      "UserOut": {
        "type": "object",
        "description": "The user shape every auth endpoint returns.\n\n`email_verified` lets the client decide whether to prompt;\n`verification_required` tells it whether prompting means anything at all,\nsince a deployment with no mailer cannot require proof it cannot request.\nWithout the second field the UI would nag every user of an unconfigured\ninstance about something they can never resolve.",
        "required": [
          "id",
          "email",
          "plan",
          "email_verified",
          "verification_required"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "email_verified": {
            "type": "boolean"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "plan": {
            "type": "string"
          },
          "verification_required": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}